Refactored file names to address PR review comments and removed unnecessary code
