--- app/javascript/src/components/Dashboard/Notes/Pane/Create.jsx
@@ -1,29 +1,0 @@
-import React from "react";
-
-import { Pane, Typography } from "neetoui";
-
-import Form from "./Form";
-
-import { NOTES_FORM_INITIAL_FORM_VALUES } from "../constants";
-
-const NewNotePane = ({ fetchNotes, showPane, setShowPane }) => {
-  const onClose = () => setShowPane(false);
-
-  return (
-    <Pane isOpen={showPane} onClose={onClose}>
-      <Pane.Header>
-        <Typography style="h2" weight="semibold">
-          Add New Note
-        </Typography>
-      </Pane.Header>
-      <Form
-        isEdit={false}
-        note={NOTES_FORM_INITIAL_FORM_VALUES}
-        refetch={fetchNotes}
-        onClose={onClose}
-      />
-    </Pane>
-  );
-};
-
-export default NewNotePane;
